+++ s2gos-server/src/s2gos_server/services/base/__init__.py
@@ -1,0 +1,11 @@
+#  Copyright (c) 2025 by ESA DTE-S2GOS team and contributors
+#  Permissions are hereby granted under the terms of the Apache 2.0 License:
+#  https://opensource.org/license/apache-2-0.
+
+from .service_base import ServiceBase
+from .service_base import DEFAULT_CONFORMANCES
+
+__all__ = [
+    "ServiceBase",
+    "DEFAULT_CONFORMANCES",
+]